Visual Button/GUI Size
This is purely a visual change that allows for a more dynamical change to the tkinter windows size. Whatever the size of the user's screen, the changes makes the size of the master window 50% the height and 25% the width of their screen. Then for the individual button widths, they are between 25%-30% of the master window size.

This should hopefully make the button layout and display more user friendly and readable.